fix agregar producto , consultar producto add modificar producto wip Busquedas de consular proucto, reserva nueva

+++ BibliotecaClasesNetframework/ModelosODT/ProductoDTO.cs
@@ -14,8 +14,8 @@ namespace BibliotecaClasesNetframework.ModelosODT
         public int StockMinimo { get; set; }
         
         public int? IDCategoria  { get; set; }
-        public int PrecioCompra { get; set; }
-        public int PrecioVenta { get; set; }
+        public decimal PrecioCompra { get; set; }
+        public decimal PrecioVenta { get; set; }
          public string NombreCategoria { get; set; }
          public string FirstProveedor { get; set; }
         public List<ProveedorDTO> proveedores { get; set; }
